Added version 1.1.453

+++ chromium/extensions/common/stack_frame.h
@@ -17,8 +17,8 @@ namespace extensions {
 struct StackFrame {
   StackFrame();
   StackFrame(const StackFrame& frame);
-  StackFrame(size_t line_number,
-             size_t column_number,
+  StackFrame(uint32_t line_number,
+             uint32_t column_number,
              const base::string16& source,
              const base::string16& function);
   ~StackFrame();
@@ -29,8 +29,11 @@ struct StackFrame {
 
   bool operator==(const StackFrame& rhs) const;
 
-  size_t line_number;
-  size_t column_number;
+  // Note: we use uint32_t instead of size_t because this struct is sent over
+  // IPC which could span 32 & 64 bit processes. This is fine since line numbers
+  // and column numbers shouldn't exceed UINT32_MAX even on 64 bit builds.
+  uint32_t line_number;
+  uint32_t column_number;
   base::string16 source;
   base::string16 function;  // optional
 };
